Convert entire model from float64 to float32 for faster processing
Updated all numpy operations throughout the model to use float32 instead of the default float64:
- embeddings.py: Convert embedding matrix to float32
- positional_encoding.py: Convert positional encodings to float32
- feed_forward.py: Convert all weights and biases to float32
- single_head_attention.py: Convert attention weight matrices and masks to float32
- transformer_block.py: Convert layer normalization parameters to float32
- output_layer.py: Convert output bias to float32
- loss_function.py: Update docstrings to reflect float32 usage

This change reduces memory usage and improves computational performance.

+++ src/embeddings/embeddings.py
@@ -36,7 +36,7 @@ class EmbeddingLayer:
         self.max_seq_length = max_seq_length
         self.n = n
 
-        self.embeddings = np.random.randn(self.vocab_size, self.embedding_dim) * np.sqrt(1.0/self.vocab_size) # Basically, random numbers are selected for the vectors right now as placeholder so that the algorithm doesn't see symmetry and simply assign the same vector values to every word upon training
+        self.embeddings = (np.random.randn(self.vocab_size, self.embedding_dim) * np.sqrt(1.0/self.vocab_size)).astype(np.float32) # Basically, random numbers are selected for the vectors right now as placeholder so that the algorithm doesn't see symmetry and simply assign the same vector values to every word upon training
 
         self.positional_encoding_class = PositionalEncoding(self.embedding_dim, self.max_seq_length)
